+++ bin/miseen.js
@@ -6,6 +6,7 @@ const chalk = require( "chalk" );
 // imports: lib
 const miseen = require( "../lib" );
 const displayError = require( "../lib/helpers/displayError" );
+const { IS_VALID_PERMISSION_MODE } = require( "../lib/helpers/regexTests" );
 
 const { info } = console;
 
@@ -17,31 +18,36 @@ const { info } = console;
  * @param {string} actionName
  * @return {{ default: string | string[] } | { [key: string]: string | string[] } | null}
  */
-const validateAction = ( action, actionName ) => {
-  let actionObject = {};
+const extraValidation = ( item, arrayOfProps ) => {
+  if ( arrayOfProps.length === 0 ) return true;
+  return arrayOfProps.filter( ( prop ) => {
+    return prop.regex.test( item[prop.name] );
+  } ).length > 0;
+};
 
+const validateAction = ( action, actionName, propsValidation = [] ) => {
   if ( Array.isArray( action ) ) {
-    actionObject = {
-      default: action
-    };
-  } else if ( typeof action === "string" ) {
-    actionObject = {
-      default: [action]
-    };
-  } else if ( typeof action === "object" ) {
-    const isInvalidMkdirCollection = Object.keys( action ).map( ( key ) => typeof action[key] === "string" || Array.isArray( action[key] ) ).includes( false );
-    if ( isInvalidMkdirCollection ) {
-      displayError( actionName + " is invalid" );
-    } else {
-      actionObject = action;
-    }
-  } else {
-    displayError( actionName + " is invalid" );
+    const onlyValidConfigItens = action.filter( ( item, index ) => {
+      const isValidItem = item.path && typeof item.path === "string" && extraValidation( item, propsValidation );
+      if ( !isValidItem ) {
+        info(
+          chalk.gray( "* (" + actionName + "[" + index + "]) " + JSON.stringify( item ) + " is not a valid config" )
+        );
+        if ( !item.path ) {
+          info(
+            chalk.gray( "    * item.path undefined" )
+          );
+        }
+      }
+      return isValidItem;
+    } );
+
+    if ( onlyValidConfigItens.length === 0 ) return null;
+
+    return onlyValidConfigItens;
   }
 
-  if ( Object.keys( actionObject ).length === 0 ) return null;
-
-  return actionObject;
+  return null;
 };
 
 /**
@@ -55,7 +61,7 @@ const getConfig = ( config ) => {
   let validConfig = {};
 
   if ( config.chmod ) {
-    validConfig.chmod = validateAction( config.chmod, "config.chmod" );
+    validConfig.chmod = validateAction( config.chmod, "config.chmod", [{ name: "mode", regex: IS_VALID_PERMISSION_MODE }] );
   }
 
   if ( config.mkdir ) {
@@ -70,6 +76,7 @@ const getConfig = ( config ) => {
 */
 
 const runScript = () => {
+  const scriptParams = process.argv;
   // Script start message
   info( chalk.gray( "> mise en package" ) );
 
@@ -82,7 +89,7 @@ const runScript = () => {
       }
 
       const config = getConfig( result.config );
-      miseen( config );
+      miseen( config, scriptParams );
     } )
     .catch( ( error ) => {
       info( chalk.red( error ) );