+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+## 0.2.0
+> **Publishing date:** 2022-11-05
+
+- **Action:** chmod
+- **Refactor:** Actions params
+
 ## 0.1.0
 > **Publishing date:** 2022-11-05
 
 - Alpha release 🎉
-- **Feat:** mkdir
+- **Action:** mkdir